fix DTO Staff,Custommer. + testOrder

+++ SWP391/src/main/java/com/ADNService/SWP391/repository/ServiceRepository.java
@@ -3,6 +3,6 @@ package com.ADNService.SWP391.repository;
 import com.ADNService.SWP391.entity.Services;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<Services, Integer> {
+public interface ServiceRepository extends JpaRepository<Services, Long> {
 
 }
